Fixed frame setup + version agnosticism
Copying frame parameters broke tty frames, and the latest seems to
copy colours without trouble. We set the frame font in the frame's
property alist, which causes new frames to respect the font. Even
better, setting the font this way is benign in emacs22, so we're no
longer emacs23 dependent.